backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
python-multipart = "^0.0.20"
coqui-tts = "^0.26.2"
openai-whisper = "^20240930"
sqlite-utils = "^3.38"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
